Move LDSConnection back to v2
